+++ shardis-auth/src/main/java/com/shardis/auth/controllers/rest/AuthRestController.java
@@ -1,0 +1,19 @@
+package com.shardis.auth.controllers.rest;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.security.Principal;
+
+/**
+ * Created by Tomasz Kucharzyk
+ */
+@RestController
+@RequestMapping("/")
+public class AuthRestController {
+
+    @RequestMapping("/user")
+    public Principal user(Principal user) {
+        return user;
+    }
+}