+++ shardis-ui/src/main/frontend/vendor.ts
@@ -1,0 +1,17 @@
+// Angular 2
+import 'angular2/platform/browser';
+import 'angular2/core';
+import 'angular2/http';
+import 'angular2/router';
+
+// RxJS
+import 'rxjs/Rx';
+
+if ('production' === ENV) {
+  // Production
+
+
+} else {
+  // Development
+
+}